requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# Testing and data
pytest>=7.4.0
pytest-asyncio>=0.21.0
faker>=20.0.0

# Utilities
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0
